+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-- N/A
+- Added benchmarking script for comparing different methods (`#87`, `!59`)
+- Added heating method based on demandlib's BDEW method (`#70`, `!63`)
 
 ### Changed
 - Changed internal naming of core methods and parameters for improved clarity (`#90`, `!62`)